confirmacion para borrado de usuario

+++ front/src/app/components/home/configuracion/usuarios/usuarios.component.ts
@@ -3,6 +3,7 @@ import { PostService } from 'src/app/services/post.service';
 import { GetService } from '../../../../services/get.service';
 import { Usuario } from '../../../../models/usuarios.model';
 import { ToastServiceService } from 'src/app/services/toast-service.service';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'app-usuarios',
@@ -15,28 +16,31 @@ export class UsuariosComponent implements OnInit {
   usuariosTodos: Usuario[];
   permisos = [];
   cargando: boolean;
-
+  idUserAEliminar = -1;
   usuarioSeleccionado: Usuario;
-
   page: number;
   pageSize: number;
   collectionSize: number;
+  disabledForm: boolean;
 
   constructor(
     private getService: GetService,
     private postService: PostService,
-    public toastService: ToastServiceService
+    public toastService: ToastServiceService,
+    private modalService: NgbModal
     ) { }
-
 
   ngOnInit(): void {
     this.cargando = true;
+    this.disabledForm = true;
     this.page = 1;
     this.pageSize = 10;
     this.obtenerUsuarios();
   }
 
   obtenerUsuarios() : void {
+    this.cargando = true;
+    this.disabledForm = true;
     this.getService.obtenerUsuarios().subscribe(res => {
       if (res){
         this.usuariosTodos = res;
@@ -44,32 +48,59 @@ export class UsuariosComponent implements OnInit {
         this.collectionSize = res.length;
         this.refreshUsers();
         this.cargando = false;
+        this.disabledForm = false;
       } else {
         this.toastService.show('Hubo un error',{ classname: 'bg-danger text-light', delay: 2000 });
         setTimeout(() => {
           this.toastService.removeAll(),
           this.cargando = false;
+          this.disabledForm = false;
         }, 2000);
       }
     });
   }
 
-  eliminar(usuario: Usuario): void{
-    this.postService.eliminarUsuario(usuario.id).subscribe(res => {
-      if (res.Status){
-        this.toastService.show('Usuario Eliminado', { classname: 'bg-danger text-light', delay: 2000 });
-        setTimeout(() => {
-          this.toastService.removeAll()
-        }, 2000);
-        this.obtenerUsuarios();
-      } else {
-        this.toastService.show('Hubo un error',{ classname: 'bg-danger text-light', delay: 2000 });
-        setTimeout(() => {
-          this.toastService.removeAll(),
-          this.cargando = false;
-        }, 2000);
+  
+  open(content , id_dist_a_borrar): void {
+    this.idUserAEliminar = id_dist_a_borrar;
+    this.modalService.open(content, { centered: true, size: 'lg' });
+  }
+
+
+  eliminar(): void{
+    
+    if(this.idUserAEliminar != -1){
+      this.disabledForm = true;
+      this.postService.eliminarUsuario(this.idUserAEliminar).subscribe(
+        (res) => {
+          if (res.Status){
+            this.toastService.show('Usuario Eliminado', { classname: 'bg-danger text-light', delay: 2000 });
+            this.modalService.dismissAll() 
+            this.idUserAEliminar = -1
+            this.disabledForm = false;
+            this.obtenerUsuarios();
+          } else {
+            this.toastService.show('Hubo un error',{ classname: 'bg-danger text-light', delay: 2000 });
+            setTimeout(() => {
+              this.toastService.removeAll(),
+              this.idUserAEliminar = -1,
+              this.cargando = false;
+            }, 3000);
+            this.disabledForm = false;
+          }
+      },(err) =>{
+          this.toastService.show('Hubo un error',{ classname: 'bg-danger text-light', delay: 2000 });
+          setTimeout(() => {
+            this.toastService.removeAll(),
+            this.idUserAEliminar = -1,
+            this.cargando = false;
+          }, 3000);
+          this.disabledForm = false;
       }
-    });
+
+      
+      );
+    }
   }
 
   refreshUsers(): void {
